Add files via upload
there is no score in this commit (but i did test it before and it works accurately). there are now a total of FOUR working questions (data sets)! i feel like my quiz is boring but i like history and decided to do it on that subject! i should add some more little silly questions in there though....