Add files via upload
i have 6 (though the last one is commented out right now to demonstrate a working score) questions as of right now! the fifth one is different because i really wanted it to have three different correct answers. i tried to figure out how to do it in the "run_quest(var1, var2, var3)" (my defined function) for half an hour but it wasnt working out for me so i decided to dump the whole loop in for that question instead. i kind of like it.